move duplicate function into utils
