+++ verifiers/utils/process_utils.py
@@ -1,6 +1,22 @@
 """Process lifecycle utilities."""
 
+import signal
+import sys
 from multiprocessing.process import BaseProcess
+
+
+def request_parent_death_signal() -> None:
+    """Ask the Linux kernel to send SIGTERM when the parent process dies."""
+    if sys.platform != "linux":
+        return
+    try:
+        import ctypes
+
+        libc = ctypes.CDLL("libc.so.6", use_errno=True)
+        PR_SET_PDEATHSIG = 1
+        libc.prctl(PR_SET_PDEATHSIG, signal.SIGTERM)
+    except Exception:
+        pass
 
 
 def terminate_process(
